auto-advance to next song, add pagination

+++ main/screen/screen_render.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdint.h>
 
 #include "esp_bt_defs.h"
@@ -18,8 +19,8 @@ void screen_render_bt_scan(const screen_bt_device_t* bt_devices,
                            int max_bt_devices,
                            uint8_t num_bt_devices,
                            int selected_bt_device_idx);
-void screen_render_song_selection(int selected_song_idx);
-void screen_render_song_loading(int selected_song_idx);
-void screen_render_song_playing(int selected_song_idx);
+void screen_render_song_selection(size_t selected_song_idx);
+void screen_render_song_loading(size_t selected_song_idx);
+void screen_render_song_playing(size_t selected_song_idx);
 void screen_render_song_paused(bool paused);
 void screen_render_song_progress(uint32_t elapsed_seconds, uint32_t total_seconds);
